requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
openai
pydantic